Added raw (Fudge-C) type constructos to message, string, date and time. Will be required until FudgeField usage is replaced with a C++ wrapper class. Added FieldFunctions and IntegerFieldDowncasting unit test implementations to Message test module.

+++ tests/simpletest.cpp
@@ -181,7 +181,7 @@ void Test::equalsTrue ( const char * file, int line, const char * text, int valu
     log ( ) << file << ":" << line << " : " << text << " = true" << std::endl;
 }
 
-void Test::equalsInt ( const char * file, int line, const char * xStr, const char * yStr, int x, int y )
+void Test::equalsInt ( const char * file, int line, const char * xStr, const char * yStr, int64_t x, int64_t y )
 {
     if ( x != y )
     {
